% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateCNV.R
\name{annotateCNV}
\alias{annotateCNV}
\title{Annotate a CNV Call}
\usage{
annotateCNV(chr, start, end, type, num_copies, reference)
}
\arguments{
\item{chr}{An integer of the chromosome number that CNV
region is in.}

\item{start}{An integer of the starting positon (bp) of
the CNV region.}

\item{end}{An integer or a character string of the ending positon (bp) of
the CNV region.}

\item{type}{A character string of either "DEL" or "DUP" indicating deletion
or duplication CNV.}

\item{num_copies}{An integer of the number of copies of the CNV. Note that
the baseline is 2 (for diploid). So num_copies = 3 would indicate 1 extra
copy (duplication), and num_copies = 1 would indicate 1 less copy
(deletion).}

\item{reference}{A character string of either "GRCh37" or "GRCh38" indicating
the reference genome that the annotation will be based on.}
}
\value{
Returns a table where each row represents 1 annotated gene, and the
   columns contain the gene name (symbol), chromosome, start position, end
   position, type of CNV that it belongs to, and the number of copy changes
   of that CNV.
}
\description{
Given a CNV call, which will include the chromosome number, start & end
position, type of CNV (DUP/DEL), and the number of copies, return all genes
that are fully contained in this region based on the GRCh37 or GRCh38
reference genome. This function will return a list of genes that are fully
encompassed by the input CNV regions, whether each gene is duplicated or
deleted, and the number of copies.
}
\examples{
# Example 1:
# One CNV input.
annotatedResult <- annotateCNV(chr = 1, start = 15654424, end = 15680097,
                               type = 'DEL', num_copies = 1,
                               reference = 'GRCh37')
# Example 2:
# Using the sample input file of 200 CNV calls, may take longer to run.
\dontrun{
load('~/CNVds/data/sampleInputCNV.rda')
annotated <- NULL
for (i in seq_along(1:nrow(sampleInputCNV))) {
  output <- annotateCNV(sampleInputCNV[i, 1], sampleInputCNV[i, 2],
                        sampleInputCNV[i, 3], sampleInputCNV[i, 4],
                        sampleInputCNV[i, 5], reference = 'GRCh37')
  annotated <- rbind(annotated, output)
}
}
}
