% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDSscores.R
\name{findpHI}
\alias{findpHI}
\title{Find pHI Scores for a List of Genes}
\usage{
findpHI(lstGenes)
}
\arguments{
\item{lstGenes}{A character vector of gene names.}
}
\value{
Returns a table of size (length of lstGenes) x 2, where each row
contains the gene name and its pHI value (if found).
}
\description{
Given a list of genes, find the corresponding pHI scores for each gene.
Probability of haploinsufficiency (pHI) is the probability that a gene is sensitive to copy number loss.
}
\examples{
# Example 1:
# Using the annotations from the output of the annotateCNV() function.
annotatedResult <- annotateCNV(chr = 1, start = 15654424, end = 15680097,
                               type = 'DEL', num_copies = 1,
                               reference = 'GRCh37')
genepHI <- findpHI(annotatedResult$gene)

# Example 2:
# Using a user-defined vector of strings consisting of gene names.
genepHI <- findpHI(c('ABHD15', 'ABLIM2', 'APOC1'))

}
\references{
Collins, R. L. et al. (2022). A cross-disorder dosage sensitivity map of the human
genome. \emph{Cell} 185, 3041-3055.e25. \href{https://www.sciencedirect.com/science/article/abs/pii/S0092867422007887#sec4.1}{Link}
}
