% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDSscores.R
\name{findpLI}
\alias{findpLI}
\title{Find pLI Scores for a List of Genes}
\usage{
findpLI(lstGenes)
}
\arguments{
\item{lstGenes}{A character vector of gene names.}
}
\value{
Returns a table of size (length of lstGenes) x 2, where each row
contains the gene name and its pLI value (if found).
}
\description{
Given a list of genes, find the corresponding pLI scores for each gene.
Probability of loss intolerance (pLI) is the probability that a gene is intolerant to a loss of function mutation.
}
\examples{
# Example 1:
# Using the annotations from the output of the annotateCNV() function.
annotatedResult <- annotateCNV(chr = 1, start = 15654424, end = 15680097,
                               type = 'DEL', num_copies = 1,
                               reference = 'GRCh37')
genepLI <- findpLI(annotatedResult$gene)

# Example 2:
# Using a user-defined vector of strings consisting of gene names.
genepLI <- findpLI(c('ABHD5', 'ABLIM2', 'PRAMEF17'))

}
\references{
Karczewski, K. J. et al. (2020). The mutational constraint spectrum
quantified from variation in 141,456 humans.
\emph{Nature} 581, 434–443. \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7334197/}{Link}
}
