% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDSscores.R
\name{geneDSscores}
\alias{geneDSscores}
\title{Find pLI, pHI, pTS Scores of a Gene}
\usage{
geneDSscores(gene)
}
\arguments{
\item{gene}{A character string that represents a gene symbol.}
}
\value{
Returns a 3x1 table containing the gene's pLI, pHI, and pTS values.
}
\description{
Given a single gene symbol, find its associated pLI, pHI, and pTS scores.
}
\examples{
geneDSscores <- geneDSscores('CHD8')

}
