% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37}
\alias{grch37}
\title{GRCh37 Subset}
\format{
A matrix with columns:
\describe{
\item{symbol}{Gene symbol, which is a short-form abbreviation for a
particular gene.}
\item{chr}{The chromosome which the gene is on.}
\item{start}{The starting position (in bp) of the gene.}
\item{end}{The ending position (in bp) of the gene.}
}
}
\source{
The dataset is directly obtained from the 'annotables' package
(https://github.com/stephenturner/annotables). GRCh37 (Genome
Reference Consortium Human Build 37) are published by Ensembl
(https://useast.ensembl.org/info/about/publications.html).
}
\usage{
grch37
}
\description{
A subset of the GRCh37 reference genome dataset to only contain the gene symbol,
chromosome number, and the start & end position of all autosomal genes.
}
\examples{
\dontrun{
 grch37
}
}
\keyword{datasets}
