% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch38}
\alias{grch38}
\title{GRCh38 Subset}
\format{
A matrix with columns:
\describe{
 \item{symbol}{Gene symbol, which is a short-form abbreviation for a
   particular gene.}
 \item{chr}{The chromosome which the gene is on.}
 \item{start}{The starting position (in bp) of the gene.}
 \item{end}{The ending position (in bp) of the gene.}
}
}
\source{
The dataset is directly obtained from the 'annotables' package
(https://github.com/stephenturner/annotables). GRCh38 (Genome
Reference Consortium Human Build 38) are published by Ensembl
(https://useast.ensembl.org/info/about/publications.html).
}
\usage{
grch38
}
\description{
A subset of the GRCh38 reference genome dataset to only contain gene symbol,
chromosome, and the start & end position of each gene.
}
\examples{
\dontrun{
 grch38
}
}
\keyword{datasets}
