% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pLI_data}
\alias{pLI_data}
\title{Probability of Loss Intolerance (pLI) Reference}
\format{
A matrix with columns:
\describe{
 \item{gene}{Gene symbol, which is a short-form abbreviation for a
   particular gene.}
 \item{pLI}{Probability of loss intolerance.}
}
}
\source{
Karczewski, K. J. et al. The mutational constraint spectrum
quantified from variation in 141,456 humans. \emph{Nature} 581, 434–443
(2020). \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7334197/}{Link}
}
\usage{
pLI_data
}
\description{
A table containing 19658 autosomal genes and their corresponding probability
of loss intolerance, which is the probability that a gene is intolerant to a
loss of function mutation.
}
\examples{
\dontrun{
 pLI_data
}
}
\keyword{datasets}
