% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoresByChr.R
\name{plotScoresByChr}
\alias{plotScoresByChr}
\title{Plot Dosage Sensitivity Scores of the Input Genes Grouped by Chromosomes}
\usage{
plotScoresByChr(genes, DStype, thresh)
}
\arguments{
\item{genes}{A table consisting of the following columns, in order:
chromosome, gene symbol, score.}

\item{DStype}{A character string of the type of DS scores.}

\item{thresh}{A floating point number between 0 and 1 that serves as the
threshold (default is 0.80). Any genes with scores above the threshold
will be labeled.}
}
\value{
Returns a combined scatter plot (one for each chromosome) of the
   distribution of the gene scores.
}
\description{
Given a list of genes and their corresponding dosage sensitivity scores (pLI/pHI/pTS), plot
the distribution of the scores and label genes that are especially
dosage-sensitive (i.e. above a certain user-defined threshold).
}
\details{
Dosage sensitivity score metrics:
Probability of loss intolerance (pLI) is the probability that a gene is intolerant to a loss of function mutation.
Probability of haploinsufficiency (pHI) is the probability that a gene is sensitive to copy number loss.
Probability of triplosensitivity (pTS) is the probability that the gene is sensitive to copy number gain.
}
\examples{
# Example 1:
# Using the sample input file of 200 CNV calls, may take longer to run.

inputPath <- system.file("extdata", "sampleInputCNV.csv", package = "CNVds")
# Read data
sampleInputCNV <- read.csv(file = inputPath, header = TRUE)

annotated <- NULL
for (i in seq_along(1:nrow(sampleInputCNV))) {
  output <- annotateCNV(sampleInputCNV[i, 1], sampleInputCNV[i, 2],
                        sampleInputCNV[i, 3], sampleInputCNV[i, 4],
                        sampleInputCNV[i, 5], reference = 'GRCh37')
  annotated <- rbind(annotated, output)
}

pHIscores <- findpHI(annotated$gene)
annotated <- merge(annotated, pHIscores, by='gene')
plotScoresByChr(annotated[c('chr', 'gene', 'pHI')], 'pHI', 0.8)


# Example 2:
# Using a small dataset of 4 random data (not real genes & scores).
annotated <- data.frame(c(1, 4, 6, 7),
                        c('geneA', 'geneB', 'geneC', 'geneD'),
                        c(0.7, 0.99, 0.2, 0.5))
plotScoresByChr(annotated, 'pHI', 0.8)

}
