% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleInputCNV}
\alias{sampleInputCNV}
\title{Sample Input of CNV Calls}
\format{
A matrix with columns:
\describe{
 \item{chr}{The chromosome which the gene is on.}
 \item{start}{The starting position (in bp) of the gene.}
 \item{end}{The ending position (in bp) of the gene.}
 \item{type}{Deletion ('DEL') or duplication ('DUP') CNV.}
 \item{num_copies}{Number of copies of the CNV.}
}
}
\source{
Ming, C, Wang, M, Wang, Q, et al. Whole genome sequencing–based copy
number variations reveal novel pathways and targets in Alzheimer's disease.
\emph{Alzheimer's & Dementia} 18, 1846- 1867 (2022).
\href{https://doi.org/10.1002/alz.12507}{Link}
}
\usage{
sampleInputCNV
}
\description{
A sample dataset of 200 CNV calls.
}
\examples{
\dontrun{
 sampleInputCNV
}
}
\keyword{datasets}
