% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDSscores.R
\name{geneDSscores}
\alias{geneDSscores}
\title{Find pLI, pHI, pTS Scores of a Gene}
\usage{
geneDSscores(gene)
}
\arguments{
\item{gene}{A character string that represents a gene symbol.}
}
\value{
Returns a 3x1 table containing the gene's pLI, pHI, and pTS values.
}
\description{
Given a single gene symbol, find its associated pLI, pHI, and pTS scores.
}
\details{
Dosage sensitivity score metrics:
Probability of loss intolerance (pLI) is the probability that a gene is intolerant to a loss of function mutation.
Probability of haploinsufficiency (pHI) is the probability that a gene is sensitive to copy number loss.
Probability of triplosensitivity (pTS) is the probability that the gene is sensitive to copy number gain.
}
\examples{
geneDSscores <- geneDSscores('CHD8')

}
