% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesNoScores.R
\name{genesNoScores}
\alias{genesNoScores}
\title{Return Genes With No Dosage Sensitivity Scores Available}
\usage{
genesNoScores(lstGenes, typeScore)
}
\arguments{
\item{lstGenes}{A character vector of gene names.}

\item{typeScore}{A character string indicating the type of dosage sensitivity
scores ('pLI'/'pHI'/'pTS').}
}
\value{
Returns list of genes that don't have a score in the corresponding
   reference table, and print out the percent of genes without a score.
}
\description{
Given a list of genes and the name of the DS score, return the list of genes
that don't have a score in the corresponding reference table, and print
out the percent of genes without a score.
}
\examples{

# Example 1:
# Using the annotations from the output of the annotateCNV() function.
annotatedResult <- annotateCNV(chr = 1, start = 15654424, end = 15680097,
                               type = 'DEL', num_copies = 1,
                               reference = 'GRCh37')
genesWithoutScores <- genesNoScores(annotatedResult$symbol, 'pTS')

# Example 2:
# Using a user-defined vector of strings consisting of gene names.
genesWithoutScores <- genesNoScores(c('ABHD15', 'ABLIM2', 'APOC1',
'randomstring'), 'pTS')

}
