% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pHaplo_pTriplo_data}
\alias{pHaplo_pTriplo_data}
\title{Probability of Haploinsufficiency/Triplosensitivity (pHI/pTS) Reference}
\format{
A matrix with columns:
\describe{
 \item{X.Gene}{Gene symbol, which is a short-form abbreviation for a
   particular gene.}
 \item{pHaplo}{Probability of haploinsufficiency.}
 \item{pTriplo}{Probability of triplosensitivity.}
}
}
\source{
Collins, R. L. et al. A cross-disorder dosage sensitivity map of the
human genome. \emph{Cell} 185, 3041-3055.e25 (2022). \href{https://www.sciencedirect.com/science/article/abs/pii/S0092867422007887#sec4.1}{Link}
}
\usage{
pHaplo_pTriplo_data
}
\description{
A table containing 18641 autosomal genes and their corresponding probability
of haploinsuffciency/triplosensitivity scores.
}
\examples{
\dontrun{
 pHaplo_pTriplo_data
}
}
\keyword{datasets}
