% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVcounts.R
\name{plotCNVcounts}
\alias{plotCNVcounts}
\title{Plot Number of CNVS Grouped by Chromosomes}
\usage{
plotCNVcounts(lstCNV)
}
\arguments{
\item{lstCNV}{}
}
\value{
Returns a barplot of the number of duplication and deletion CNVs
   grouped by chromosomes.
}
\description{
Given a list of CNV calls, plot the total number of duplication and deletion
CNVs in each chromosome.
}
