% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVcounts.R
\name{plotCNVcounts}
\alias{plotCNVcounts}
\title{Plot Number of CNVs Grouped by Chromosomes}
\usage{
plotCNVcounts(lstCNV)
}
\arguments{
\item{lstCNV}{A table consisting of the following columns, in order:
chromosome number, start position, end position,
type of CNV ('DUP'/'DEL').}
this item values for (9556455906) ;
}
\value{
Returns a barplot of the number of duplication and deletion CNVs
grouped by chromosomes.
}
\description{
Given a list of copy number variation (CNV) calls, plot the total number of unique duplication and
deletion CNV regions in each chromosome.
}
\examples{
# Example 1:
# A mini example of 4 random CNV regions.
CNVinput <- data.frame(c(1, 1, 2, 3),
                       c(12312, 293012, 3829103, 3281901),
                       c(23423, 402405, 4025002, 8531273),
                       c('DEL', 'DUP', 'DEL', 'DEL'))
plotCNVcounts(CNVinput)

# Example 2:
# Using the sample input of a list of CNV calls.
inputPath <- system.file("extdata", "sampleInputCNV.csv", package = "CNVds")
# Read data
sampleInputCNV <- read.csv(file = inputPath, header = TRUE)
plotCNVcounts(sampleInputCNV[c(1:4)])

}
