% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCNVds.R
\name{runCNVds}
\alias{runCNVds}
\title{Launch Shiny App for CNVds}
\usage{
runCNVds()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for CNVds.
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

CNVds::runCNVds()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
